requirements.txt
toolz
scikit-image
pandas

pytorch-ignite==0.2.0
pytorch-pretrained-bert==0.6.2
torch==1.1.0
torchtext==0.3.1
torchvision==0.3.0
Pillow==5.4.1

efficientnet-pytorch

